PDFBOX-4071: Sonar fix: when expecting a long one of the parameters needs to be a long
git-svn-id: https://svn.apache.org/repos/asf/pdfbox/trunk@1827699 13f79535-47bb-0310-9956-ffa450edef68

+++ pdfbox/src/main/java/org/apache/pdfbox/pdmodel/graphics/image/SampledImageReader.java
@@ -363,9 +363,9 @@ final class SampledImageReader
             {
                 // we just need to copy all sample data, then convert to RGB image.
                 long inputResult = input.read(bank);
-                if (Long.compare(inputResult, width * height * (int) numComponents) != 0)
+                if (Long.compare(inputResult, width * height * (long) numComponents) != 0)
                 {
-                    LOG.debug("Tried reading " + width * height * numComponents + " bytes but only " + inputResult + " bytes read");
+                    LOG.debug("Tried reading " + width * height * (long) numComponents + " bytes but only " + inputResult + " bytes read");
                 }
                 return pdImage.getColorSpace().toRGBImage(raster);
             }
